Build a dataset from derived

This function builds a dataset out of the columns that just need to be pulled through. So any variable that has a derivation in the format of 'dataset.variable' will be pulled through to create the new dataset. These columns are often called 'Predecessors' in ADaM, but this is not universal so that is optional to specify.

build_from_derived(
  metacore,
  ds_list,
  dataset_name = NULL,
  predecessor_only = FALSE
)

Arguments

metacore

Metacore object that contains the specifications for the dataset of interest.

ds_list

Named list of datsets that are needed to build the from

dataset_name

Optional string to specify the dataset. This is only needed if the metacore object provided hasn't already been subsetted.

predecessor_only

By default `FALSE`, but if `TRUE` will only use derivations with the origin of 'Predecessor'

Value

datset

Examples

library(metacore)
library(haven)
library(magrittr)
metacore <- define_to_metacore(metacore_example("ADaM_define.xml"), quiet = TRUE) %>%
  select_dataset("ADSL")
#> 
#>  Metadata successfully imported
#> Loading in metacore object with suppressed warnings
ds_list <- list(DM = read_xpt(metatools_example("dm.xpt")))
build_from_derived(metacore, ds_list)
#> # A tibble: 306 × 13
#>    STUDYID     USUBJID SUBJID SITEID TRT01P   AGE AGEU  RACE  SEX   ETHNIC DTHFL
#>    <chr>       <chr>   <chr>  <chr>  <chr>  <dbl> <chr> <chr> <chr> <chr>  <chr>
#>  1 CDISCPILOT… 01-701… 1015   701    Place…    63 YEARS WHITE F     HISPA… ""   
#>  2 CDISCPILOT… 01-701… 1023   701    Place…    64 YEARS WHITE M     HISPA… ""   
#>  3 CDISCPILOT… 01-701… 1028   701    Xanom…    71 YEARS WHITE M     NOT H… ""   
#>  4 CDISCPILOT… 01-701… 1033   701    Xanom…    74 YEARS WHITE M     NOT H… ""   
#>  5 CDISCPILOT… 01-701… 1034   701    Xanom…    77 YEARS WHITE F     NOT H… ""   
#>  6 CDISCPILOT… 01-701… 1047   701    Place…    85 YEARS WHITE F     NOT H… ""   
#>  7 CDISCPILOT… 01-701… 1057   701    Scree…    59 YEARS WHITE F     HISPA… ""   
#>  8 CDISCPILOT… 01-701… 1097   701    Xanom…    68 YEARS WHITE M     NOT H… ""   
#>  9 CDISCPILOT… 01-701… 1111   701    Xanom…    81 YEARS WHITE F     NOT H… ""   
#> 10 CDISCPILOT… 01-701… 1115   701    Xanom…    84 YEARS WHITE M     NOT H… ""   
#> # … with 296 more rows, and 2 more variables: RFSTDTC <chr>, RFENDTC <chr>