Check Variable Names

Usage

check_variables(data, metacore, dataset_name = deprecated(), strict = TRUE)

Arguments

data

Dataset to check

metacore

metacore object that only contains the specifications for the dataset of interest.

dataset_name

Optional string to specify the dataset. This is only needed if the metacore object provided hasn't already been subsetted.
Note: Deprecated in version 1.0.0. The dataset_name argument will be removed in a future release. Please use metacore::select_dataset to subset the metacore object to obtain metadata for a single dataset.

strict

A logical value indicating whether to perform strict validation on the input dataset. If TRUE (default), errors will be raised if validation fails. If FALSE, warnings will be issued instead, allowing the function execution to continue event with invalid data.

message if the dataset matches the specification and the dataset, and error otherwise This function checks the variables in the dataset against the variables defined in the metacore specifications. If everything matches the function will print a message stating everything is as expected. If there are additional or missing variables an error will explain the discrepancies library(haven) library(metacore) library(magrittr) load(metacore_example("pilot_ADaM.rda")) spec <- metacore %>% select_dataset("ADSL") data <- read_xpt(metatools_example("adsl.xpt")) check_variables(data, spec) data["DUMMY_COL"] <- NA check_variables(data, spec, strict = FALSE)