Convert Variable to Factor with Levels Set by Control Terms

This functions takes a dataset, a metacore object and a variable name. Then looks at the metacore object for the control terms for the given variable and uses that to convert the variable to a factor with those levels. If the control terminology is a code list, the code column will be used. The function fails if the control terminology is an external library

convert_var_to_fct(data, metacore, var)

Arguments

data

A dataset containing the variable to be modified

metacore

A metacore object to get the codelist from. If the variable has different codelists for different datasets the metacore object will need to be subsetted using `select_dataset` from the metacore package

var

Name of variable to change

Value

Dataset with variable changed to a factor

Examples

library(metacore)
library(haven)
library(magrittr)
spec <- spec_to_metacore(metacore_example("p21_mock.xlsx"), quiet = TRUE)
#> 
#>  Metadata successfully imported
#> Loading in metacore object with suppressed warnings
dm <- read_xpt(metatools_example("dm.xpt"))
# Variable with codelist control terms
convert_var_to_fct(dm, spec, SEX)
#> # A tibble: 306 × 25
#>    STUDYID      DOMAIN USUBJID  SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>        <chr>  <chr>    <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCPILOT01 DM     01-701-… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCPILOT01 DM     01-701-… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCPILOT01 DM     01-701-… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCPILOT01 DM     01-701-… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCPILOT01 DM     01-701-… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCPILOT01 DM     01-701-… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCPILOT01 DM     01-701-… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCPILOT01 DM     01-701-… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCPILOT01 DM     01-701-… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCPILOT01 DM     01-701-… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 16 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <fct>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>
# Variable with permitted value control terms
convert_var_to_fct(dm, spec, ARM)
#> # A tibble: 306 × 25
#>    STUDYID      DOMAIN USUBJID  SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>        <chr>  <chr>    <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCPILOT01 DM     01-701-… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCPILOT01 DM     01-701-… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCPILOT01 DM     01-701-… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCPILOT01 DM     01-701-… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCPILOT01 DM     01-701-… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCPILOT01 DM     01-701-… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCPILOT01 DM     01-701-… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCPILOT01 DM     01-701-… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCPILOT01 DM     01-701-… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCPILOT01 DM     01-701-… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 16 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <fct>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>