Create Categorical Variable from Codelist

Using the grouping from either the `decode_var` or `code_var` and a reference variable (`ref_var`) it will create a categorical variable and the numeric version of that categorical variable.

create_cat_var(data, metacore, ref_var, grp_var, num_grp_var = NULL)

Arguments

data

Dataset with reference variable in it

metacore

A metacore object to get the codelist from. If the variable has different codelists for different datasets the metacore object will need to be subsetted using `select_dataset` from the metacore package.

ref_var

Name of variable to be used as the reference i.e AGE when creating AGEGR1

grp_var

Name of the new grouped variable

num_grp_var

Name of the new numeric decode for the grouped variable. This is optional if no value given no variable will be created

Value

dataset with new column added

Examples

library(metacore)
library(haven)
library(magrittr)
spec <- define_to_metacore(metacore_example("ADaM_define.xml"), quiet = TRUE) %>%
  select_dataset("ADSL")
#> 
#>  Metadata successfully imported
#> Loading in metacore object with suppressed warnings
dm <- read_xpt(metatools_example("dm.xpt"))
# Grouping Column Only
create_cat_var(dm, spec, AGE, AGEGR1)
#> # A tibble: 306 × 26
#>    STUDYID      DOMAIN USUBJID  SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>        <chr>  <chr>    <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCPILOT01 DM     01-701-… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCPILOT01 DM     01-701-… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCPILOT01 DM     01-701-… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCPILOT01 DM     01-701-… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCPILOT01 DM     01-701-… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCPILOT01 DM     01-701-… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCPILOT01 DM     01-701-… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCPILOT01 DM     01-701-… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCPILOT01 DM     01-701-… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCPILOT01 DM     01-701-… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 17 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>, AGEGR1 <chr>
# Grouping Column and Numeric Decode
create_cat_var(dm, spec, AGE, AGEGR1, AGEGR1N)
#> # A tibble: 306 × 27
#>    STUDYID      DOMAIN USUBJID  SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>        <chr>  <chr>    <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCPILOT01 DM     01-701-… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCPILOT01 DM     01-701-… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCPILOT01 DM     01-701-… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCPILOT01 DM     01-701-… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCPILOT01 DM     01-701-… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCPILOT01 DM     01-701-… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCPILOT01 DM     01-701-… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCPILOT01 DM     01-701-… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCPILOT01 DM     01-701-… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCPILOT01 DM     01-701-… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 18 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>, AGEGR1 <chr>, AGEGR1N <chr>